+++ libs/ledger/tests/chaincode/smart_contract_tests.cpp
@@ -18,10 +18,10 @@
 
 #include "contract_test.hpp"
 #include "core/containers/is_in.hpp"
+#include "core/string/replace.hpp"
 #include "crypto/ecdsa.hpp"
 #include "crypto/sha256.hpp"
 #include "ledger/chain/transaction_builder.hpp"
-//#include "ledger/chain/transaction.hpp"
 #include "ledger/chaincode/smart_contract.hpp"
 #include "ledger/state_adapter.hpp"
 #include "mock_storage_unit.hpp"
@@ -38,6 +38,7 @@ using ::testing::Return;
 
 using fetch::byte_array::ConstByteArray;
 using fetch::core::IsIn;
+using fetch::string::Replace;
 using fetch::ledger::Address;
 using fetch::ledger::SmartContract;
 using fetch::storage::ResourceAddress;
@@ -568,20 +569,6 @@ TEST_F(SmartContractTests, CheckShardedStateSetWithAddressAsName)
   auto request{Variant::Object()};
   request["address"] = address_as_name.display();
   VerifyQuery("query_foo", int32_t{20}, request);
-}
-
-bool Replace(std::string &orig, std::string const &what, std::string const &with)
-{
-  auto start{orig.find(what)};
-
-  if (start == std::string::npos)
-  {
-    return false;
-  }
-
-  orig.replace(start, what.size(), with);
-
-  return true;
 }
 
 TEST_F(SmartContractTests, CheckContextInAction)
